agents_runtime/requirements.txt
fastapi>=0.115,<1
uvicorn>=0.30,<1
langchain>=0.3,<0.4
langchain-openai>=0.2,<0.3
python-dotenv>=1.0,<2
pydantic>=2.8,<3
